ScreenSwitcher displays prefabs based on model's current scene
