+++ Boccia-Unity/Assets/Boccia/UI/Menu Prefabs/TempStartMenu.prefab
@@ -541,6 +541,7 @@ GameObject:
   - component: {fileID: 7198735677603073068}
   - component: {fileID: 7828267550846357144}
   - component: {fileID: 6127433248351135154}
+  - component: {fileID: 2260097816163904203}
   m_Layer: 5
   m_Name: StartMenu
   m_TagString: Untagged
@@ -633,6 +634,20 @@ MonoBehaviour:
   m_BlockingMask:
     serializedVersion: 2
     m_Bits: 4294967295
+--- !u!114 &2260097816163904203
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3266750391644024403}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e2afd72019fc143cf931d401c69abbc4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  startButton: {fileID: 6890300007834131507}
+  quitButton: {fileID: 8292551770593260641}
 --- !u!1 &5052558058256109230
 GameObject:
   m_ObjectHideFlags: 0
